refactor(cli): delete resolve-config.ts and loadClientFromPath
Delete resolve-config.ts (209 lines, zero callers). Collapse
loadClientFromPath into loadConfig—workspace commands now call
loadConfig(dir) and pick clients[0]. Flatten the default-vs-named
export distinction: Object.entries(module) already includes the
default export, so no special-casing needed. Move DiscoveredWorkspace
type to workspace-command.ts (its only consumer). Remove dead exports
from index.ts (resolveWorkspace, discoverWorkspaces, WorkspaceResolution).
